Actually compile the test
